Move Page class to pages folder
