+++ store/helpers/pages/index.js
@@ -1,0 +1,1 @@
+export Pages from './pages'
